# server will populate the compute and queue times for ensemble models