Instead of just printing 0 for compute and queue times for ubrella ensemble, server will set them to the total sum of the compute and queue times of the composing models.