@tanmayv25 2019-08-13
The outputs that demonstrate the effect of PR:

### Non - Ensemble:
PR does not affect the output
```

 Request concurrency: 3
   Client: 
     Request count: 266
     Throughput: 88 infer/sec
     Avg latency: 33807 usec (standard deviation 4257 usec)
     p50 latency: 33675 usec
     p90 latency: 39793 usec
     p95 latency: 41108 usec
     p99 latency: 44551 usec
     Avg HTTP time: 33520 usec (send/recv 651 usec + response wait 32869 usec)
   Server: 
     Request count: 321
     Avg request latency: 31188 usec (overhead 16 usec + queue 20043 usec + compute 11129 usec)
```

### Ensemble (The details of the model used is at the end):
#### Before:

```
 *** Measurement Settings ***
   Batch size: 1
   Measurement window: 3000 msec
   Latency limit: 50 msec
   Concurrency limit: 1 concurrent requests
   Stabilizing using average latency
 
 Request concurrency: 1
   Client: 
     Request count: 4043
     Throughput: 1347 infer/sec
     Avg latency: 728 usec (standard deviation 111 usec)
    p50 latency: 702 usec
     p90 latency: 857 usec
     p95 latency: 924 usec
     p99 latency: 1114 usec
     Avg HTTP time: 670 usec (send/recv 39 usec + response wait 631 usec)
   Server: 
     Request count: 4854
     Avg request latency: 449 usec (overhead 449 usec + queue 0 usec + compute 0 usec)
 
   Composing models: 
   simple, version: 1
     Request count: 4853
     Avg request latency: 197 usec (overhead 4 usec + queue 19 usec + compute 174 usec)

   simple, version: 2
     Request count: 4854
     Avg request latency: 192 usec (overhead 5 usec + queue 18 usec + compute 169 usec)
 
 Inferences/Second vs. Client Average Batch Latency
 Concurrency: 1, 1347 infer/sec, latency 728 usec
```

#### After:

```
 *** Measurement Settings ***
   Batch size: 1
   Measurement window: 3000 msec
   Latency limit: 50 msec
   Concurrency limit: 1 concurrent requests
   Stabilizing using average latency
 
 Request concurrency: 1
   Client: 
     Request count: 4052
     Throughput: 1350 infer/sec
     Avg latency: 727 usec (standard deviation 111 usec)
     p50 latency: 701 usec
     p90 latency: 862 usec
     p95 latency: 931 usec
     p99 latency: 1080 usec
     Avg HTTP time: 666 usec (send/recv 40 usec + response wait 626 usec)
   Server: 
     Request count: 4903
     Avg request latency: 443 usec
     Total avg compute time : 340 usec
     Total avg queue time : 35 usec
 
   Composing models: 
   simple, version: 1
     Request count: 4903
     Avg request latency: 192 usec (overhead 5 usec + queue 17 usec + compute 170 usec)
 
   simple, version: 2
     Request count: 4903
     Avg request latency: 192 usec (overhead 6 usec + queue 17 usec + compute 169 usec)
 
 Inferences/Second vs. Client Average Batch Latency
 Concurrency: 1, 1350 infer/sec, latency 727 usec


```
#### ModelConfig
```
 ensemble_scheduling {
   step [
     {
       model_name: "simple"
       model_version: 1
       input_map {
         key: "INPUT0"
         value: "INPUT0"
       }
       input_map {
         key: "INPUT1"
         value: "INPUT1"
       }
       output_map {
         key: "OUTPUT0"
         value: "temp0"
       }
       output_map {
         key: "OUTPUT1"
         value: "temp1"
       }
     },
     {
       model_name: "simple"
       model_version: 2
       input_map {
         key: "INPUT0"
         value: "temp0"
       }
       input_map {
         key: "INPUT1"
         value: "temp1"
       }
       output_map {
         key: "OUTPUT0"
         value: "OUTPUT0"
       }
       output_map {
         key: "OUTPUT1"
         value: "OUTPUT1"
       }
     }
   ]
 }
 
```