@GuanLuo 2019-08-13
Since the total compute / queue time will be reported, should this be done in the server side? Say in the server status, the compute / queue time of an ensemble model will simply be the total time aggregated from all composing models.

One advantage of this is that the data will be accurate in the case of nested ensemble. For example, if one of the composing model is another ensemble model, the compute / queue time of it will just be 0. In the extreme case where all composing models are ensembles, the perf_client will report 0 for both total compute time and total queue time. Unless the perf_client will rollout the nested ensemble models and aggregate the their stats first, which is naturally been done in ensemble scheduler.