@tanmayv25 2019-08-14
After these changes on server side, we can populate the compute and queue times for the nested ensembles as well. 

#### Before
```

*** Measurement Settings ***
  Batch size: 1
  Measurement window: 3000 msec
  Latency limit: 50 msec
  Concurrency limit: 1 concurrent requests
  Stabilizing using average latency

Request concurrency: 1
  Client: 
    Request count: 2508
    Throughput: 836 infer/sec
    Avg latency: 1182 usec (standard deviation 118 usec)
    p50 latency: 1169 usec
    p90 latency: 1306 usec
    p95 latency: 1388 usec
    p99 latency: 1540 usec
    Avg HTTP time: 1113 usec (send/recv 36 usec + response wait 1077 usec)
  Server: 
    Request count: 3015
    Avg request latency: 879 usec
    Total avg compute time : 0 usec
    Total avg queue time : 0 usec

  Composing models: 
  ensemble, version: 1
    Request count: 3016
    Avg request latency: 424 usec (overhead 424 usec + queue 0 usec + compute 0 usec)

  ensemble, version: 2
    Request count: 3015
    Avg request latency: 394 usec (overhead 394 usec + queue 0 usec + compute 0 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, 836 infer/sec, latency 1182 usec
```

#### After

```
*** Measurement Settings ***
  Batch size: 1
  Measurement window: 3000 msec
  Latency limit: 50 msec
  Concurrency limit: 1 concurrent requests
  Stabilizing using average latency

Request concurrency: 1
  Client: 
    Request count: 2509
    Throughput: 836 infer/sec
    Avg latency: 1181 usec (standard deviation 167 usec)
    p50 latency: 1148 usec
    p90 latency: 1407 usec
    p95 latency: 1486 usec
    p99 latency: 1693 usec
    Avg HTTP time: 1118 usec (send/recv 41 usec + response wait 1077 usec)
  Server: 
    Request count: 3009
    Avg request latency: 881 usec
    Total avg compute time : 653 usec
    Total avg queue time : 66 usec

  Composing models: 
  ensemble, version: 1
    Request count: 3008
    Avg request latency: 427 usec (overhead 53 usec + queue 35 usec + compute 339 usec)

  ensemble, version: 2
    Request count: 3009
    Avg request latency: 394 usec (overhead 49 usec + queue 31 usec + compute 314 usec)
```
#### Model Config
```

ensemble_scheduling {
  step [
    {
      model_name: "ensemble"
      model_version: 1
      input_map {
        key: "INPUT0"
        value: "INPUT0"
      }
      input_map {
        key: "INPUT1"
        value: "INPUT1"
      }
      output_map {
        key: "OUTPUT0"
        value: "temp0"
      }
      output_map {
        key: "OUTPUT1"
        value: "temp1"
      }
    },
    {
      model_name: "ensemble"
      model_version: 2
      input_map {
        key: "INPUT0"
        value: "temp0"
      }
      input_map {
        key: "INPUT1"
        value: "temp1"
      }
      output_map {
        key: "OUTPUT0"
        value: "OUTPUT0"
      }
      output_map {
        key: "OUTPUT1"
        value: "OUTPUT1"
      }
    }
  ]

```