@deadeyegoodwin 2019-08-14
When the composing model is an ensemble shouldn't it be printed in the ensemble style (i.e. separate lines for  total avg compute, total avg queue, etc. As it is now won't it be confusing when queue or compute time is greater than total latency?